No macros, no cry!

+++ trunk/mtl4/trunk/boost/numeric/mtl/detail/contiguous_memory_block.hpp
@@ -55,7 +55,7 @@ struct size_helper
     void set_size(std::size_t size)
     {
 #     ifndef MTL_IGNORE_STATIC_SIZE_VIOLATION
-	MTL_THROW_IF(Size != size, change_static_size());
+	throw_if(Size != size, change_static_size());
 #     endif
     }
 
@@ -274,7 +274,7 @@ struct contiguous_memory_block
     void copy_assignment(const Other& other)
     {
 	// std::cout << "Copied in assignment.\n";	
-	MTL_DEBUG_THROW_IF(this->used_memory() != other.used_memory(), incompatible_size());
+	debug_throw_if(this->used_memory() != other.used_memory(), incompatible_size());
 	std::copy(other.data, other.data + other.used_memory(), data);
     }
 
@@ -366,7 +366,7 @@ public:
 	// If already have memory of the right size we can keep it
 	if (size == this->used_memory()) 
 	    return;
-	MTL_THROW_IF(category != own, 
+	throw_if(category != own, 
 		     logic_error("Can't change the size of collections with external memory"));
 	delete_it();
 	alloc(size);
@@ -402,7 +402,7 @@ struct contiguous_memory_block<Value, true, Size>
     Value    data[Size];
     explicit contiguous_memory_block(std::size_t size= Size) 
     {
-	MTL_DEBUG_THROW_IF(Size != size, incompatible_size());
+	debug_throw_if(Size != size, incompatible_size());
     }
 
     // Move-semantics ignored for arrays on stack
@@ -417,7 +417,7 @@ struct contiguous_memory_block<Value, true, Size>
     explicit contiguous_memory_block(const contiguous_memory_block<Value2, OnStack2, Size2>& other)
     {
 	// std::cout << "Copied in copy constructor (different type).\n";	
-	MTL_DEBUG_THROW_IF(Size != other.used_memory(), incompatible_size());
+	debug_throw_if(Size != other.used_memory(), incompatible_size());
 	std::copy(other.data, other.data + other.used_memory(), data);
     }
 
@@ -440,7 +440,7 @@ public:
     self& operator=(const contiguous_memory_block<Value2, OnStack2, Size2>& other)
     {
 	// std::cout << "Assignment from different type.\n";
-	MTL_DEBUG_THROW_IF(Size != other.used_memory(), incompatible_size());
+	debug_throw_if(Size != other.used_memory(), incompatible_size());
 	std::copy(other.data, other.data + other.used_memory(), data);
 	return *this;
     }
